+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/package.json
@@ -13,7 +13,6 @@
 	"devDependencies": {
 		"@inlang/paraglide-js": "workspace:*",
 		"@inlang/paraglide-js-adapter-sveltekit": "workspace:*",
-		"@inlang/paraglide-js-adapter-vite": "workspace:*",
 		"@inlang/plugin-message-format": "workspace:*",
 		"@sveltejs/adapter-static": "^2.0.3",
 		"@sveltejs/kit": "^1.20.4",
