+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/src/routes/[[lang=lang]]/about/+page.svelte
@@ -1,1 +1,3 @@
+<script></script>
+
 <h1>About</h1>
